refactor: 멤버십 관련 이벤트 처리 로직 멤버십 핸들러로 이동
